Rename toy_tables_macro to font_types_macro
