+++ font-types-macro/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "toy-table-macro"
+name = "font-types-macro"
 version = "0.0.0"
 edition = "2021"
 license = "MIT/Apache-2.0"
