+++ font-types-macro/tests/bad-attrs.rs
@@ -1,4 +1,4 @@
-toy_table_macro::tables! {
+font_types_macro::tables! {
     TooMuchCount<'a> {
         item_count: Uint16,
         #[count(item_count)]
@@ -11,7 +11,7 @@ fn double(val: font_types::BigEndian<u16>) -> usize {
     val.get() as usize * 2
 }
 
-toy_table_macro::tables! {
+font_types_macro::tables! {
     BadCountFn<'a> {
         item_count: Uint16,
         #[count_with(double, Self::item_count)]
@@ -19,14 +19,14 @@ toy_table_macro::tables! {
     }
 }
 
-toy_table_macro::tables! {
+font_types_macro::tables! {
     MissingCount<'a> {
         item_count: Uint24,
         items: [Uint24],
     }
 }
 
-toy_table_macro::tables! {
+font_types_macro::tables! {
     CountOnScalar<'a> {
         #[count(item_count)]
         item_count: Uint24,
@@ -35,7 +35,7 @@ toy_table_macro::tables! {
     }
 }
 
-toy_table_macro::tables! {
+font_types_macro::tables! {
     HiddenOnArray<'a> {
         item_count: Uint24,
         #[hidden]
