+++ font-types-macro/tests/count-all-guard.rs
@@ -1,6 +1,6 @@
 use font_types::BigEndian;
 
-toy_table_macro::tables! {
+font_types_macro::tables! {
     CountAll<'a> {
         item_count: BigEndian<u16>,
         #[count_all]
