+++ font-types-macro/tests/count-all-guard.stderr
@@ -1,7 +1,7 @@
 error: #[count_all] annotation only valid on last field (TODO: validate before here)
   --> tests/count-all-guard.rs:3:1
    |
-3  | / toy_table_macro::tables! {
+3  | / font_types_macro::tables! {
 4  | |     CountAll<'a> {
 5  | |         item_count: BigEndian<u16>,
 6  | |         #[count_all]
@@ -10,4 +10,4 @@ error: #[count_all] annotation only valid on last field (TODO: validate before h
 10 | | }
    | |_^
    |
-   = note: this error originates in the macro `toy_table_macro::tables` (in Nightly builds, run with -Z macro-backtrace for more info)
+   = note: this error originates in the macro `font_types_macro::tables` (in Nightly builds, run with -Z macro-backtrace for more info)
