+++ font-types-macro/tests/enums.rs
@@ -1,54 +1,43 @@
-use font_types::{BigEndian, Version16Dot16};
+use font_types::{Version16Dot16, BigEndian};
 
 const VERSION_0_5: Version16Dot16 = Version16Dot16::new(0, 5);
 const VERSION_1_0: Version16Dot16 = Version16Dot16::new(1, 0);
 
-toy_table_macro::tables! {
+mod consts {
+    pub const ONE: u16 = 1;
+}
+
+font_types_macro::tables! {
     Maxp05 {
          version: BigEndian<Version16Dot16>,
-         teeth: BigEndian<u16>,
+         num_glyphs: BigEndian<u16>,
     }
 
     Maxp10 {
          version: BigEndian<Version16Dot16>,
          num_glyphs: BigEndian<u16>,
+         max_points: BigEndian<u16>,
+         max_contours: BigEndian<u16>,
+         max_composite_points: BigEndian<u16>,
     }
 
     #[format(Version16Dot16)]
     enum Maxp {
-        // missing version
+        #[version(VERSION_0_5)]
         Version0_5(Maxp05),
         #[version(VERSION_1_0)]
         Version1_0(Maxp10),
     }
 }
 
-toy_table_macro::tables! {
-    One {
-         one: BigEndian<u16>,
-    }
-
-    Two {
-         two: BigEndian<u16>,
-    }
-
-    // missing version format
-    enum OneOrTwo {
-        #[version(VERSION_0_5)]
-        One(One),
-        #[version(VERSION_0_5)]
-        Two(Two),
-    }
-}
-
-toy_table_macro::tables! {
+font_types_macro::tables! {
     One {
          one: BigEndian<u16>,
     }
 
     #[format(u16)]
     enum OneOrTwo {
-        #[version(MISSING_VERSION)]
+        #[version(consts::ONE)]
         One(One),
     }
 }
