+++ font-types-macro/tests/smoke-test.rs
@@ -1,6 +1,6 @@
 use font_types::{BigEndian, F2Dot14};
 
-toy_table_macro::tables! {
+font_types_macro::tables! {
     SegmentMaps<'a> {
         position_map_count: BigEndian<u16>,
         #[count(position_map_count)]
